+++ src/main/java/com/vm/service/QuestionService.java
@@ -6,7 +6,7 @@ import com.vm.request.QuestionObject;
 import java.util.List;
 
 public interface QuestionService {
-	public List<QuestionObject> getQuestionBySurveyId(Long survey_id);
-	public List<NewQuestionObject> getQuestionWithNewFormatBySurveyId(Long survey_id);
+	public List<QuestionObject> getQuestionBySurveyId(Integer survey_id);
+	public List<NewQuestionObject> getQuestionWithNewFormatBySurveyId(Integer survey_id);
 	public QuestionObject getQuestionById(Long question_id);
 }
